Move RecursiveSortFilterProxyModel to tools.models + fix imports
