+++ avalon/tools/projectmanager/model.py
@@ -1,5 +1,3 @@
-import re
-
 from ...vendor.Qt import QtCore
 
 
@@ -28,32 +26,3 @@ class ExactMatchesFilterProxyModel(QtCore.QSortFilterProxyModel):
                 return True
             else:
                 return False
-
-
-class RecursiveSortFilterProxyModel(QtCore.QSortFilterProxyModel):
-    """Filters to the regex if any of the children matches allow parent"""
-    def filterAcceptsRow(self, row, parent):
-
-        regex = self.filterRegExp()
-        if not regex.isEmpty():
-            pattern = regex.pattern()
-            model = self.sourceModel()
-            source_index = model.index(row, self.filterKeyColumn(), parent)
-            if source_index.isValid():
-
-                # Check current index itself
-                key = model.data(source_index, self.filterRole())
-                if re.search(pattern, key, re.IGNORECASE):
-                    return True
-
-                # Check children
-                rows = model.rowCount(source_index)
-                for i in range(rows):
-                    if self.filterAcceptsRow(i, source_index):
-                        return True
-
-                # Otherwise filter it
-                return False
-
-        return super(RecursiveSortFilterProxyModel,
-                     self).filterAcceptsRow(row, parent)
